SimpleAgent/requirements.txt
openai
python-dotenv
pyyaml
colorama
requests
beautifulsoup4
googlesearch-python
PyGithub
discord.py